image_reader/requirements.txt
asyncio-nats-client==0.11.4
aiounittest==1.4.0
